pyproject.toml
[project]
name = "stt-simple"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faster-whisper>=1.2.1",
    "ffmpeg-python>=0.2.0",
    "numpy>=2.4.2",
    "ruff>=0.15.0",
    "scipy>=1.17.0",
    "setuptools-rust>=1.12.0",
    "ty>=0.0.15",
]

[project.optional-dependencies]
dev = ["pyinstaller>=6.0.0"]
